+++ test/numerical_checks.jl
@@ -27,7 +27,7 @@ using ManifoldsBaseTestUtils
             limits = (-2.5, 2.0),
             error = :error,
         )
-        @test !check_retraction(M, ProjectionRetraction(), p, X)
+        @test !check_retraction(M, ProjectionRetraction(), p, X; limits = (-2.5, 2.0))
 
         #test window size error
         @test_throws ErrorException ManifoldsBase.find_best_slope_window(
@@ -44,6 +44,64 @@ using ManifoldsBaseTestUtils
         check_retraction(
             M,
             ExponentialRetraction(),
+            p,
+            X;
+            plot = true,
+            exactness_tol = 1e-7,
+        )
+    end
+    @testset "Test Inverse Retraction checks" begin
+        M = TestSphere(10)
+        q = zeros(11)
+        q[1] = 1.0
+        p = zeros(11)
+        p[1:4] .= 1 / sqrt(4)
+        X = log(M, p, q)
+
+        @test check_inverse_retraction(
+            M,
+            ProjectionInverseRetraction(),
+            p,
+            X;
+            limits = (-2.5, 0.0),
+        )
+        # One call with generating a plot
+        check_inverse_retraction(
+            M,
+            ProjectionInverseRetraction(),
+            p,
+            X;
+            limits = (-2.5, 0.0),
+            plot = true,
+        )
+
+        # ProjectionRetraction only works <= 1 well in stepsize
+        @test_throws ErrorException check_inverse_retraction(
+            M,
+            ProjectionInverseRetraction(),
+            p,
+            X;
+            limits = (-2.5, 2.0), # yields a bit too long tangents
+            error = :error,
+        )
+        @test !check_inverse_retraction(
+            M,
+            ProjectionInverseRetraction(),
+            p,
+            X;
+            limits = (-2.5, 2.0),
+        )
+
+        @test check_inverse_retraction(
+            M,
+            LogarithmicInverseRetraction(),
+            p,
+            X;
+            exactness_tol = 1e-7,
+        )
+        check_inverse_retraction(
+            M,
+            LogarithmicInverseRetraction(),
             p,
             X;
             plot = true,